Remove scheduled scan from SCA
The scheduled scan and notification was introduced on our main repo as
we control all of the code and we wanted to have notification if there
were SCA issues that needed to be address.  Since this repo has shared
code of both MP and teams, a sheduled run offers little value as we are
not enforcing SCA standards on users.  This would mean we get daily
notifications when there are issues that we are not going to look at.
Therefore removing the scheduled run from this workflow.